Refactor sponsorship components and enhance layout for better user experience
